refactor(012): extract global-styles + parameterize table enhancement (US3)
Move the ~95-line global CSS literal out of bootstrap.ts into init/global-styles.ts
and collapse the duplicated enhanceAllQuizTables/enhanceAllAnalysisTables helpers
into one parameterized enhanceAllTables. bootstrap.ts is now 321 lines.
